requirements.txt
ccxt
pandas
python-dotenv
ta-lib
fastapi
uvicorn
websockets
jinja2
python-multipart
feedparser
textblob
aiohttp
